refactor: split api-websockets-server abstractions into feature folders
Each abstraction (ServerAdapter, UpgradeHandler, ConnectionManager) now
lives in its own folder's abstractions.ts instead of a single monolithic file.
